Backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.2
botocore==1.37.2
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
filelock==3.17.0

# FastAPI dependencies (replacing Flask)
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
websockets==14.1

fonttools==4.56.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
openai==1.63.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.3
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
termcolor==2.5.0
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==2.3.0

# Core dependencies
python-dotenv==1.0.1

# OpenAI
openai
# async-openai==0.18.2

# Database
SQLAlchemy==2.0.25

# HTML/Feed Processing
# beautifulsoup4==4.12.3

# Async Support
aiohttp==3.9.3
asyncio==3.4.3

# Logging & Monitoring
# termcolor==2.4.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

#
redis==5.2.1
jsonpath-ng==1.4.3
bitsandbytes
accelerate

# Deferred processing
celery
google-generativeai